docs: RelatedComponents - fix <Dialog> description
