+++ src/storybook/components/related-components/descriptions/dialog-description/dialog-description.module.scss
@@ -1,0 +1,4 @@
+.container {
+  width: 100%;
+  padding-left: var(--spacing-large);
+}
